Introduce `safeAttribute(_:_:_:)`
The `isSafeAttribute(_:_:_:)` had the unexpected side-effect of changing the attribute's value to turn it from a relative to an absolute URL, if so configured. But this was modifying the ORIGINAL attribute, and thus broke the promise of `Cleaner` to not modify its source document.

So I moved this code to a new method that clones the attribute and does the modification on the clone instead, if necessary.